+++ server/services/integrations/registry.js
@@ -12,6 +12,7 @@ const { createSlackProvider } = require('./slack/provider');
 const { createWeatherProvider } = require('./weather/provider');
 const { createWhatsAppPersonalProvider } = require('./whatsapp');
 const { createNeoArchiveProvider } = require('./neoarchive/provider');
+const { createNeoRecallProvider } = require('./neorecall/provider');
 
 function createIntegrationRegistry(options = {}) {
   const providers = [
@@ -25,6 +26,7 @@ function createIntegrationRegistry(options = {}) {
     createWeatherProvider(),
     createSpotifyProvider(),
     createNeoArchiveProvider(),
+    createNeoRecallProvider(),
     createHomeAssistantProvider(),
     createWhatsAppPersonalProvider(options),
   ];